milestone4/backend/requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
langchain==0.1.20
langchain-openai
langchain-community
reportlab